requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
alembic
httpx
pydantic-settings
fastapi-cache2
fake-useragent